The data type determining the value of the product has been changed from int to BigDecimal. Added a filter by price and product name

+++ src/main/java/ru/geelbrains/spring/winter/market/dtos/CartDto.java
@@ -3,6 +3,8 @@ package ru.geelbrains.spring.winter.market.dtos;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,6 +12,6 @@ import java.util.List;
 @AllArgsConstructor
 public class CartDto {
     private List<CartItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 }
