+++ src/main/java/ru/geelbrains/spring/winter/market/entities/Product.java
@@ -5,6 +5,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -25,7 +27,7 @@ public class Product {
     private Category category;
 
     @Column(name = "price")
-    private int price;
+    private BigDecimal price;
 
     @Column(name = "height")
     private int height;
